+++ src/main/java/com/zea7ot/leetcode/logs.txt
@@ -4,7 +4,7 @@
 5: 
 
 20200906:
-3: 1080
+3: 0147, 1080
 4: 
 5:
